fix index corruption, broodlings, overseers

+++ src/cost.py
@@ -4,15 +4,15 @@ from dataclasses import dataclass
 @dataclass
 class Cost:
 
-    minerals: int
-    vespene: int
-    food: int
+    minerals: float
+    vespene: float
+    food: float
     larva: float
 
     def __add__(self, other: 'Cost'):
         return Cost(self.minerals + other.minerals, self.vespene + other.vespene, self.food + other.food, self.larva + other.larva)
 
-    def __mul__(self, factor: int):
+    def __mul__(self, factor: float):
         return Cost(self.minerals * factor, self.vespene * factor, self.food * factor, self.larva * factor)
 
     def __repr__(self) -> str:
